Added history list. Down arrow shows history items when there is no search text. Navigating suggestions when no other text has been entered will autopopulate the input box. Windows: Plugins and skins can be installed in <profile>\Application Data\Launchy\Skins and Plugins respectively. Very basic fallback skin built into the executable. Windows: Icons up to 48x48 supported if the outputIcon skin element is larger than 32x32. Fixed warning message Tweaks to make Verby less aggressive

+++ Launchy_QT/src/Fader.cpp
@@ -41,7 +41,7 @@ void Fader::fadeIn(bool quick)
 	mutex.lock();
 	targetLevel = gSettings->value("GenOps/opaqueness", 100).toInt() / 100.0;
 	delta = 0.05;
-	delay = quick ? 0 : time * delta / targetLevel;
+	delay = quick ? 0 : (int)(time * delta / targetLevel);
 	if (delay > 10)
 	{
 		delta /= 10;
@@ -72,7 +72,7 @@ void Fader::fadeOut(bool quick)
 	mutex.lock();
 	targetLevel = 0;
 	delta = -0.05;
-	delay = quick ? 0 : time * -delta / opaqueness;
+	delay = quick ? 0 : (int)(time * -delta / opaqueness);
 	if (delay > 10)
 	{
 		delta /= 10;
